+++ main.go
@@ -18,215 +18,14 @@ package main
 
 import (
 	"fmt"
-	"github.com/urfave/cli"
-	"log"
-	"net/http"
-	"net/url"
 	"os"
-	"time"
-
-	"github.com/MohawkTSDB/mohawk/backend"
-	"github.com/MohawkTSDB/mohawk/backend/example"
-	"github.com/MohawkTSDB/mohawk/backend/memory"
-	"github.com/MohawkTSDB/mohawk/backend/mongo"
-	"github.com/MohawkTSDB/mohawk/backend/sqlite"
-	"github.com/MohawkTSDB/mohawk/middleware"
-	"github.com/MohawkTSDB/mohawk/router"
+	"github.com/MohawkTSDB/mohawk/cmd"
 )
 
-// VER the server version
-const VER = "0.21.4"
-
-// defaults
-const defaultAPI = "0.21.0"
-const defaultTLSKey = "server.key"
-const defaultTLSCert = "server.pem"
-
-// BackendName Mohawk active backend
-var BackendName string
-
-// GetStatus return a json status struct
-func GetStatus(w http.ResponseWriter, r *http.Request, argv map[string]string) {
-	resTemplate := `{"MetricsService":"STARTED","Implementation-Version":"%s","MohawkVersion":"%s","MohawkBackend":"%s"}`
-	res := fmt.Sprintf(resTemplate, defaultAPI, VER, BackendName)
-
-	w.WriteHeader(200)
-	fmt.Fprintln(w, res)
-}
-
 func main() {
-	app := cli.NewApp()
-	app.Name = "mohawk"
-	app.Version = VER
-	app.Usage = "Metric data storage engine"
-	app.Authors = []cli.Author{
-		{Name: "Yaacov Zamir", Email: "kobi.zamir@gmail.com"},
+	if err := cmd.RootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
 	}
-	app.Flags = []cli.Flag{
-		cli.StringFlag{Name: "backend,b", Value: "memory", Usage: "the backend plugin to use"},
-		cli.StringFlag{Name: "token", Value: "", Usage: "authorization token"},
-		cli.StringFlag{Name: "media", Value: "./mohawk-webui", Usage: "path to media files"},
-		cli.StringFlag{Name: "key", Value: defaultTLSKey, Usage: "path to TLS key file"},
-		cli.StringFlag{Name: "cert", Value: defaultTLSCert, Usage: "path to TLS cert file"},
-		cli.StringFlag{Name: "options", Value: "", Usage: "specific backend options [e.g. db-dirname, db-url]"},
-		cli.UintFlag{Name: "port,p", Value: 8080, Usage: "server port"},
-		cli.BoolFlag{Name: "tls,t", Usage: "use TLS server"},
-		cli.BoolFlag{Name: "gzip,g", Usage: "enable gzip encoding"},
-		cli.BoolFlag{Name: "verbose,V", Usage: "more debug output"},
-	}
-	app.Action = serve
-	app.Run(os.Args)
 }
 
-func serve(c *cli.Context) error {
-	var db backend.Backend
-	var middlewareList []middleware.MiddleWare
-
-	// Create and init the backend
-	switch c.String("backend") {
-	case "sqlite":
-		db = &sqlite.Backend{}
-	case "memory":
-		db = &memory.Backend{}
-	case "mongo":
-		db = &mongo.Backend{}
-	case "example":
-		db = &example.Backend{}
-	default:
-		log.Fatal("Can't find backend:", c.String("backend"))
-	}
-
-	// parse options
-	if options, err := url.ParseQuery(c.String("options")); err == nil {
-		db.Open(options)
-	} else {
-		log.Fatal("Can't parse opetions:", c.String("options"))
-	}
-
-	// set global variables
-	BackendName = db.Name()
-
-	// h common variables to be used for the backend Handler functions
-	// backend the backend to use for metrics source
-	h := backend.Handler{
-		Verbose: c.Bool("verbose"),
-		Backend: db,
-	}
-
-	// Create the routers
-	// Requests not handled by the routers will be forworded to BadRequest Handler
-	rRoot := router.Router{
-		Prefix: "/hawkular/metrics/",
-	}
-	// Root Routing table
-	rRoot.Add("GET", "status", GetStatus)
-	rRoot.Add("GET", "tenants", h.GetTenants)
-	rRoot.Add("GET", "metrics", h.GetMetrics)
-
-	// Metrics Routing tables
-	rGauges := router.Router{
-		Prefix: "/hawkular/metrics/gauges/",
-	}
-	rGauges.Add("GET", ":id/raw", h.GetData)
-	rGauges.Add("GET", ":id/stats", h.GetData)
-	rGauges.Add("POST", "raw", h.PostData)
-	rGauges.Add("POST", "raw/query", h.PostQuery)
-	rGauges.Add("PUT", "tags", h.PutMultiTags)
-	rGauges.Add("PUT", ":id/tags", h.PutTags)
-	rGauges.Add("DELETE", ":id/raw", h.DeleteData)
-	rGauges.Add("DELETE", ":id/tags/:tags", h.DeleteTags)
-
-	// deprecated
-	rGauges.Add("GET", ":id/data", h.GetData)
-	rGauges.Add("POST", "data", h.PostData)
-	rGauges.Add("POST", "stats/query", h.PostQuery)
-
-	rCounters := router.Router{
-		Prefix: "/hawkular/metrics/counters/",
-	}
-	rCounters.Add("GET", ":id/raw", h.GetData)
-	rCounters.Add("GET", ":id/stats", h.GetData)
-	rCounters.Add("POST", "raw", h.PostData)
-	rCounters.Add("POST", "raw/query", h.PostQuery)
-	rCounters.Add("PUT", ":id/tags", h.PutTags)
-
-	// deprecated
-	rCounters.Add("GET", ":id/data", h.GetData)
-	rCounters.Add("POST", "data", h.PostData)
-	rCounters.Add("POST", "stats/query", h.PostQuery)
-
-	rAvailability := router.Router{
-		Prefix: "/hawkular/metrics/availability/",
-	}
-	rAvailability.Add("GET", ":id/raw", h.GetData)
-	rAvailability.Add("GET", ":id/stats", h.GetData)
-
-	// Create the middlewares
-	// logger a logging middleware
-	logger := middleware.Logger{
-		Verbose: c.Bool("verbose"),
-	}
-
-	// static a file server middleware
-	static := middleware.Static{
-		Verbose:   c.Bool("verbose"),
-		MediaPath: c.String("media"),
-	}
-
-	// authorization middleware
-	authorization := middleware.Authorization{
-		Verbose:         c.Bool("verbose"),
-		UseToken:        c.String("token") != "",
-		PublicPathRegex: "^/hawkular/metrics/status$",
-		Token:           c.String("token"),
-	}
-
-	// headers a headers middleware
-	headers := middleware.Headers{
-		Verbose: c.Bool("verbose"),
-	}
-
-	// gzipper a gzip encoding middleware
-	gzipper := middleware.GZipper{
-		UseGzip: c.Bool("gzip"),
-		Verbose: c.Bool("verbose"),
-	}
-
-	// badrequest a BadRequest middleware
-	badrequest := middleware.BadRequest{
-		Verbose: c.Bool("verbose"),
-	}
-
-	// concat middlewars and routes (first logger until rRoot) with a fallback to BadRequest
-	middlewareList = []middleware.MiddleWare{
-		&logger,
-		&authorization,
-		&headers,
-		&gzipper,
-		&rGauges,
-		&rCounters,
-		&rAvailability,
-		&rRoot,
-		&static,
-		&badrequest,
-	}
-	middleware.Append(middlewareList)
-
-	// Run the server
-	srv := &http.Server{
-		Addr:           fmt.Sprintf("0.0.0.0:%d", c.Int("port")),
-		Handler:        logger,
-		ReadTimeout:    10 * time.Second,
-		WriteTimeout:   10 * time.Second,
-		MaxHeaderBytes: 1 << 20,
-	}
-	if c.Bool("tls") {
-		log.Printf("Start server, listen on https://%+v", srv.Addr)
-		log.Fatal(srv.ListenAndServeTLS(c.String("cert"), c.String("key")))
-	} else {
-		log.Printf("Start server, listen on http://%+v", srv.Addr)
-		log.Fatal(srv.ListenAndServe())
-	}
-
-	return nil
-}
